#390 Adjusts tests for ambient light and removes redundant type casting.

+++ source/core/src/main/com/csse3200/game/services/LightService.java
@@ -33,7 +33,7 @@ public class LightService {
 	 */
 	public void renderLight() {
 		float time = ServiceLocator.getTimeService().getHour() + (float) ServiceLocator.getTimeService().getMinute() / 60;
-		float brightness = (float) (MathUtils.sin((float) ((float) Math.PI * (time-2) / 23 + 0.1)));
+		float brightness = MathUtils.sin((float) ((float) Math.PI * (time-2) / 23 + 0.1));
 		rayHandler.setAmbientLight(brightness, brightness, brightness, brightness);
 		rayHandler.setCombinedMatrix(camera);
 		rayHandler.updateAndRender();
